bug fixes + wandb integration
