+++ scripts/experiments/baseline/gen_experiment.py
@@ -19,7 +19,7 @@ base_call = f"python {PROJECT_HOME}/src/train.py -o {DATA_HOME}/output"
 # the key for each variable should match the name of the command-line
 # argument required by the script in base_call
 variables = {
-    "lr": [1e-4],
+    "epochs": [100],
 }
 
 combinations = list(itertools.product(*variables.values()))
@@ -32,7 +32,7 @@ for c in combinations:
     # and recorded in the output data by the python script
     expt_call = base_call
     for i, var in enumerate(variables.keys()):
-        expt_call += f" --{var} {c[i]}"
+        expt_call += f" --{var}={c[i]}"
     print(expt_call, file=output_file)
 
 output_file.close()
